requirements.txt
asgiref==3.4.1
bleach==4.1.0
Django==3.2.7
django-modeltranslation==0.17.3
django-summernote==0.8.20.0
django-taggit==1.5.1
packaging==21.2
Pillow==8.3.2
postgres==4.0
psycopg2-binary==2.9.1
psycopg2-pool==1.1
pyparsing==2.4.7
pytz==2021.1
six==1.16.0
sqlparse==0.4.2
webencodings==0.5.1
